[Fabric] Cleans up overrides for HostPlatformViewProps
React Native mobile recently refactored ViewProps to more easily allow host platform overrides in out of tree platforms.

This change moves the logic from WindowsViewProps.h in a bespoke location to an isomorphic location under `vnext/Microsoft.ReactNative/Fabric/platform` in HostPlatformViewProps.h, which is the name of the file Fabric expects for host platform extensions. It also moves the overrides for view flattening traits to HostPlatformViewTraitsInitializer and fixes up props parsing for WindowsViewEvents when the propIteratorSetter feature is not enabled.
