+++ vnext/Microsoft.ReactNative/Fabric/platform/react/renderer/components/view/HostPlatformViewProps.h
@@ -3,8 +3,30 @@
 
 #pragma once
 
-#include <react/components/view/windows/WindowsViewProps.h>
+#include <react/renderer/components/view/BaseViewProps.h>
+#include <react/renderer/core/PropsParserContext.h>
+#include "WindowsViewEvents.h"
 
 namespace facebook::react {
-using HostPlatformViewProps = WindowsViewProps;
+class HostPlatformViewProps : public BaseViewProps {
+ public:
+  HostPlatformViewProps() = default;
+  HostPlatformViewProps(
+      const PropsParserContext &context,
+      const HostPlatformViewProps &sourceProps,
+      const RawProps &rawProps,
+      bool shouldSetRawProps = true);
+
+  void
+  setProp(const PropsParserContext &context, RawPropsPropNameHash hash, const char *propName, RawValue const &value);
+
+  WindowsViewEvents windowsEvents{};
+  bool enableFocusRing{true};
+  bool focusable{false};
+
+  // std::optional<std::string> overflowAnchor{};
+  // std::optional<std::string> tooltip{};
+  // std::vector<HandledKeyEvent> keyDownEvents{};
+  // std::vector<HandledKeyEvent> keyUpEvents{};
+};
 } // namespace facebook::react
