--- vnext/ReactCommon/TEMP_UntilReactCommonUpdate/react/renderer/components/view/BaseViewProps.h
@@ -1,112 +1,0 @@
-/*
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-#pragma once
-
-#include <react/renderer/components/view/AccessibilityProps.h>
-#include <react/renderer/components/view/YogaStylableProps.h>
-#include <react/renderer/components/view/primitives.h>
-#include <react/renderer/core/LayoutMetrics.h>
-#include <react/renderer/core/Props.h>
-#include <react/renderer/core/PropsParserContext.h>
-#include <react/renderer/graphics/Color.h>
-#include <react/renderer/graphics/Transform.h>
-
-#include <optional>
-
-namespace facebook::react {
-
-class BaseViewProps;
-
-class BaseViewProps : public YogaStylableProps, public AccessibilityProps{
- public:
-  BaseViewProps() = default;
-  BaseViewProps(
-      const PropsParserContext &context,
-      BaseViewProps const &sourceProps,
-      RawProps const &rawProps,
-      bool shouldSetRawProps = true);
-
-  void setProp(
-      const PropsParserContext &context,
-      RawPropsPropNameHash hash,
-      const char *propName,
-      RawValue const &value);
-
-#ifdef ANDROID
-  void propsDiffMapBuffer(Props const *oldProps, MapBufferBuilder &builder)
-      const override;
-#endif
-
-#pragma mark - Props
-
-  // Color
-  Float opacity{1.0};
-  SharedColor backgroundColor{};
-
-  // Borders
-  CascadedBorderRadii borderRadii{};
-  CascadedBorderColors borderColors{};
-  CascadedBorderCurves borderCurves{}; // iOS only?
-  CascadedBorderStyles borderStyles{};
-
-  // Shadow
-  SharedColor shadowColor{};
-  Size shadowOffset{0, -3};
-  Float shadowOpacity{};
-  Float shadowRadius{3};
-
-  // Transform
-  Transform transform{};
-  BackfaceVisibility backfaceVisibility{};
-  bool shouldRasterize{};
-  std::optional<int> zIndex{};
-
-  // Events
-  PointerEventsMode pointerEvents{};
-  EdgeInsets hitSlop{};
-  bool onLayout{};
-
-  ViewEvents events{};
-
-  bool collapsable{true};
-
-  bool removeClippedSubviews{false};
-
-  LayoutConformance experimental_layoutConformance{};
-
-  Float elevation{}; /* Android-only */
-
-#ifdef ANDROID
-
-  std::optional<NativeDrawable> nativeBackground{};
-  std::optional<NativeDrawable> nativeForeground{};
-
-  bool focusable{false};
-  bool hasTVPreferredFocus{false};
-  bool needsOffscreenAlphaCompositing{false};
-  bool renderToHardwareTextureAndroid{false};
-
-#endif
-
-#pragma mark - Convenience Methods
-
-  BorderMetrics resolveBorderMetrics(LayoutMetrics const &layoutMetrics) const;
-  bool getClipsContentToBounds() const;
-
-#ifdef ANDROID
-  bool getProbablyMoreHorizontalThanVertical_DEPRECATED() const;
-#endif
-
-#pragma mark - DebugStringConvertible
-
-#if RN_DEBUG_STRING_CONVERTIBLE
-  SharedDebugStringConvertibleList getDebugProps() const override;
-#endif
-};
-
-} // namespace facebook::react
